win32 support

+++ sector-sphere/trunk/client/fscache.cpp
@@ -181,7 +181,7 @@ int Cache::insert(char* block, const std::string& path, const int64_t& offset, c
    return 0;
 }
 
-int Cache::read(const std::string& path, char* buf, const int64_t& offset, const int64_t& size)
+int64_t Cache::read(const std::string& path, char* buf, const int64_t& offset, const int64_t& size)
 {
    CGuard sg(m_Lock);
 
@@ -198,7 +198,7 @@ int Cache::read(const std::string& path, char* buf, const int64_t& offset, const
       // this condition can be improved to provide finer granularity
       if ((offset >= i->m_llOffset) && (i->m_llSize - (offset - i->m_llOffset) >= size))
       {
-         memcpy(buf, i->m_pcBlock + offset - i->m_llOffset, size);
+         memcpy(buf, i->m_pcBlock + offset - i->m_llOffset, int(size));
          i->m_llLastAccessTime = CTimer::getTime();
          // update the file's last access time; it must be equal to the block's last access time
          s->second.m_llLastAccessTime = i->m_llLastAccessTime;
